server/pyproject.toml
[tool.poetry]
name = "ai-real-estate-backend"
version = "0.1.0"
description = "Backend for AI Real Estate App"
authors = ["Erickson Figueroa <indetrd@gmail.com>"]
readme = "README.md"
packages = [{include = "server"}]  # Change this to specify the actual package location

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.2"
flask-cors = "^3.0.10"
openai = "^1.42.0"
numpy = "^1.21.2"
scikit-learn = "^1.0.2"
python-dotenv = "^1.0.1"
pyodbc = "^5.1.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
pandas = "^2.2.2"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
run = "main:app.run"
